Modularized
Subdivided mousefix into base module and subscripts that import it.
